pyproject.toml
[build-system]
requires = ["setuptools>=40.6.0"]
build-backend = "setuptools.build_meta"


[project]
name = "pybench"
dynamic = ["version", "readme"]
description = "Python interfaces for lab bench instruments"
license = {text = "EPL-2.0"}
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules"]
requires-python = ">=3.7"
authors = [
    {name = "Travis Collins", email="travis.collins@analog.com"},
]
maintainers = [
    {name = "Travis Collins", email="travis.collins@analog.com"},
    {name = "Analog Devices, Inc"}
]
dependencies = [
    'numpy >= 1.20',
    'pyvisa >= 1.11.3',
    'pyvisa-py',
    'pyserial',
    'pyusb',
    'pyyaml',
]

[tool.setuptools.dynamic]
version = {attr = "bench.__version__"}
readme = {file = "README.md", content-type = "text/markdown"}

[tool.setuptools.packages.find]
where = ["."]
include = ["bench*"]
exclude = ["tests*"]
namespaces = true

[project.scripts]
pybenchiio = "bench.cli.iiotools:cli"
pybench = "bench.cli.visatools:cli"

[project.optional-dependencies]
cli = ["click", "pyadi-iio>=0.0.17"]
web = ["fastapi[standard]", "uvicorn", "pydantic", "jinja2", "starlette"]

[project.entry-points.pytest11]
pybench = "bench.plugin"

[project.urls]
homepage = "https://tfcollins.github.io/pybench/"
documentation = "https://tfcollins.github.io/pybench/"
repository = "https://github/tfcollins/pybench/"

[tool.isort]
multi_line_output=3
include_trailing_comma="True"
force_grid_wrap=0
use_parentheses="True"
line_length=88

[tool.mypy]
ignore_missing_imports="true"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--adi-hw-map"
testpaths = [
    "tests",
]
